requirements.txt
fastapi
uvicorn
pydantic
pillow
numpy
opencv-python-headless
pytesseract
python-multipart
